جستجوی مشتری و کالا به هربخش و مدیریت فروش اضافه شده است.کتابخانه سیستم پیامکی هم اضافه شده است ولی درست اضافه نشد به کتابخانه

+++ src/main/java/com/developer/springOracle3/entity/Production.java
@@ -22,7 +22,7 @@ public class Production {
     @Column
     private String prName;
     @Column
-    private Integer meterPr;
+    private Double meterPr;
     @Transient
     private Double remainMeter;
     @Column
